backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==41.0.7
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
uritemplate==4.1.1
